code/backend/scoritoAgent/requirements.txt
python-dotenv
azure-cosmos
azure-loganalytics
azure-identity
openai
azure-core
pillow
azure-ai-agents